+++ CHANGELOG.md
@@ -35,10 +35,10 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ### Features
 * Add notification/announcement support #328
-  * `/api/v2/notifications/proposals` - Count of OPEN proposals, subcount of UPGRADE proposals
+  * `/api/v2/notifications/proposals` - List of NON-UPGRADE OPEN proposals, List of UPGRADE proposals
   * `/api/v2/notifications/upgrades` - List of expected, scheduled upgrades
   * `/api/v2/notifications/announcement` - PUT an announcement to be displayed in Explorer
-  * `/api/v2/notifications/announcement/all` - Paginated list of announcements
+  * `/api/v2/notifications/announcement/all` - Paginated list of announcements, can filter by `fromDate`
   * `/api/v2/notifications/announcement/{id}` - DELETE an existing announcement
 * Add new IBC APIs #336
   * `/api/v2/txs/ibc/chain/{ibcChain}` - txs per IBC chain id, query params supporting narrowing by channel